requirements.txt
google-cloud==0.19.0
Django==1.10.1
